requirements.txt
asttokens==3.0.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fonttools==4.55.8
giotto-ph==0.2.4
giotto-tda==0.6.2
gmpy2==2.1.5
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
igraph==0.11.8
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
mpmath==1.3.0
munkres==1.1.4
narwhals==1.24.1
nest_asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==24.3.1
platformdirs==4.3.6
plotly==6.0.0
ply==3.11
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyflagser==0.4.7
Pygments==2.19.1
pyparsing==3.2.1
PyQt5==5.15.10
PyQt5_sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
sip==6.7.12
six==1.17.0
stack_data==0.6.3
sympy==1.13.3
texttable==1.7.0
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.1
unicodedata2==16.0.0
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
widgetsnbextension==4.0.13
zipp==3.21.0
zstandard==0.23.0
